[mob] Put clip queries in same db class
