+++ mobile/lib/ui/settings/ml/ml_user_dev_screen.dart
@@ -1,7 +1,6 @@
 import "package:flutter/material.dart";
 import "package:logging/logging.dart";
 import "package:photos/core/event_bus.dart";
-import "package:photos/db/ml/clip_db.dart";
 import "package:photos/db/ml/db.dart";
 import "package:photos/events/people_changed_event.dart";
 import "package:photos/service_locator.dart";
